Changes to enhance Chisel 3 compatibility
These were all SInt/UInt constants or cast changes.  I believe these changes
should do nothing in Chisel 2 -- they're all just things that have been defined
as illegal in Chisel 3 because they were implementation defined in Chisel 2.  I
changed these to explicitly match the Chisel 2 implementation's behavior.

"make test-c" still passes.

+++ src/main/scala/tests.scala
@@ -43,11 +43,11 @@ import Node._
 object equivRecFN
 {
     def apply(expWidth: Int, sigWidth: Int, a: Bits, b: Bits) = {
-        val top4A = a(expWidth + sigWidth, expWidth + sigWidth - 3)
-        val top4B = b(expWidth + sigWidth, expWidth + sigWidth - 3)
+        val top4A = a(expWidth + sigWidth, expWidth + sigWidth - 3).toUInt
+        val top4B = b(expWidth + sigWidth, expWidth + sigWidth - 3).toUInt
         Mux((top4A(2, 0) === UInt(0)) || (top4A(2, 0) === UInt(7)),
             (top4A === top4B) && (a(sigWidth - 2, 0) === b(sigWidth - 2, 0)),
-            Mux((top4A(2, 0) === UInt(6)), (top4A === top4B), (a === b))
+            Mux((top4A(2, 0) === UInt(6)), (top4A === top4B), (a.toUInt === b.toUInt))
         )
     }
 }
